472 purge test records

+++ laravel/database/migrations/2016_09_13_090604_add_foreign_keys_to_transactions_1.php
@@ -23,8 +23,16 @@ class AddForeignKeysToTransactions1 extends Migration
         \Illuminate\Support\Facades\DB::statement("ALTER TABLE `transactions_logs` CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;");
         \Illuminate\Support\Facades\DB::statement("ALTER TABLE `transaction_change_logs` CONVERT TO CHARACTER SET utf8 COLLATE utf8_general_ci;");
         Schema::table('transactions_logs', function ($table) {
-            $table->dropForeign('transactions_logs_transaction_id_foreign');
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
+            $keyExists = \Illuminate\Support\Facades\DB::select(
+                DB::raw(
+                    'SHOW KEYS
+                    FROM transactions_logs
+                    WHERE Key_name=\'transaction_id\''
+                )
+            );
+            if (!$keyExists) {
+                $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
+            }
         });
         \Illuminate\Support\Facades\DB::statement("SET foreign_key_checks = 1;");
